Add storage.InitMySQLEngine
